DevUi: moving Logo to the shared components.
