+++ devworkflow/web/package.json
@@ -16,6 +16,7 @@
 		"vite": "^2.6.14"
 	},
 	"dependencies": {
+		"@namespacelabs.dev/webui-components": "fn:namespacelabs.dev/foundation/webui/components",
 		"classnames": "^2.3.1",
 		"date-fns": "^2.28.0",
 		"react": "^17.0.2",
