+++ webui/components/package.json
@@ -1,0 +1,13 @@
+{
+	"browserslist": "> 0.5%, last 2 versions, not dead",
+	"dependencies": {
+		"react": "^17.0.2"
+	},
+	"devDependencies": {
+		"@types/react": "^17.0.35",
+		"typescript": "4.5.4",
+		"typescript-plugin-css-modules": "^3.4.0",
+		"vite": "2.7.13"
+	},
+	"private": true
+}
